python/requirements.txt
# Core packages
numpy>=1.23
scipy>=1.9
matplotlib>=3.7
nibabel>=5.2
pydicom>=2.4
antspyx>=0.3.9
hd-bet>=2.0.1